last few errors

+++ model/src/main/scala/uk/gov/ons/addressIndex/model/server/response/AddressResponse.scala
@@ -80,7 +80,7 @@ case class AddressBySearchResponse(
   limit: Int,
   offset: Int,
   total: Long,
-  maxScore: Float
+  maxScore: Double
 )
 
 object AddressBySearchResponse {
